+++ forum-app/src/main/java/ru/tsu/hits/forum/Message/service/MessageService.java
@@ -4,9 +4,13 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.client.HttpClientErrorException;
+import ru.tsu.hits.common.dto.userDto.UserRoleDto;
+import ru.tsu.hits.common.security.JwtUserData;
+import ru.tsu.hits.common.security.exception.ForbiddenException;
 import ru.tsu.hits.forum.Category.repository.CategoryRepository;
 import ru.tsu.hits.common.dto.messageDto.CreateUpdateMessageDto;
 import ru.tsu.hits.common.dto.messageDto.MessageDto;
@@ -55,10 +59,15 @@ public class MessageService {
     }
 
     @Transactional
-    public MessageDto edit(String id, CreateUpdateMessageDto createUpdateMessageDto){
+    public MessageDto edit(String id, CreateUpdateMessageDto createUpdateMessageDto, Authentication auth){
+
         var entity  = messageRepository.findById(id)
                 .orElseThrow(() -> new HttpClientErrorException((HttpStatus.NOT_FOUND)));
 
+        var user = (JwtUserData)auth.getPrincipal();
+        if((!user.getRole().equals(UserRoleDto.ADMINISTRATOR.toString())) || (!user.getId().toString().equals(entity.getAuthor()))){
+            throw new ForbiddenException("У вас нет прав изменять это сообщение");
+        }
         entity.setText(createUpdateMessageDto.getText());
         entity.setLastModDate(new Date(System.currentTimeMillis()));
 
@@ -67,9 +76,13 @@ public class MessageService {
     }
 
     @Transactional
-    public void delete(String id){
+    public void delete(String id, Authentication auth){
         var entity  = messageRepository.findById(id)
                 .orElseThrow(() -> new HttpClientErrorException((HttpStatus.NOT_FOUND)));
+        var user = (JwtUserData)auth.getPrincipal();
+        if((!user.getRole().equals(UserRoleDto.ADMINISTRATOR.toString())) || (!user.getId().toString().equals(entity.getAuthor()))){
+            throw new ForbiddenException("У вас нет удалять это сообщение");
+        }
         messageRepository.delete(entity);
     }
